RestApi/requirements.txt
asgiref==3.3.1
Django==3.1.7
django-filter==2.4.0
djangorestframework==3.12.4
importlib-metadata==3.10.0
Markdown==3.3.4
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.7.4.3
zipp==3.4.1